requirements.txt
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
python-dotenv==1.0.1
openai==1.59.5
requests==2.32.3
pydantic==2.10.5
aiohttp==3.11.11
mcp>=1.0.0
httpx>=0.27.0

